Add more utils to utils and reorganize the functions
