+++ package.json
@@ -25,7 +25,7 @@
         "@types/bluebird": "^3.5.26",
         "@types/convict": "^4.2.1",
         "@types/debug": "^4.1.2",
-        "@types/elasticsearch": "^5.0.31",
+        "@types/elasticsearch": "^5.0.32",
         "@types/fs-extra": "^5.0.5",
         "@types/jest": "^24.0.11",
         "@types/lodash": "^4.14.123",
@@ -35,7 +35,7 @@
         "@types/socket.io": "^2.1.2",
         "@types/socket.io-client": "^1.4.32",
         "@types/uuid": "^3.4.4",
-        "eslint": "^5.15.2",
+        "eslint": "^5.15.3",
         "eslint-config-airbnb-base": "^13.1.0",
         "eslint-plugin-import": "^2.16.0",
         "fs-extra": "^7.0.1",
