+++ packages/ts-transforms/package.json
@@ -59,7 +59,7 @@
         "@types/lodash": "^4.14.123",
         "@types/valid-url": "^1.0.2",
         "@types/validator": "^10.9.0",
-        "@types/yargs": "^12.0.9"
+        "@types/yargs": "^12.0.10"
     },
     "engines": {
         "node": ">=8.0.0"
