+++ packages/utils/src/index.ts
@@ -1,7 +1,11 @@
+export * from './arrays';
 export * from './collector';
 export * from './data-entity';
+export * from './dates';
 export * from './interfaces';
 export * from './logger';
+export * from './misc';
 export * from './promises';
 export * from './errors';
+export * from './strings';
 export * from './utils';
