+++ packages/utils/test/utils-spec.ts
@@ -5,11 +5,9 @@ import {
     parseJSON,
     getTypeOf,
     isEmpty,
-    withoutNil,
     parseList,
     parseNumberList,
     toNumber,
-    toSafeString,
 } from '../src';
 
 describe('Utils', () => {
@@ -134,45 +132,6 @@ describe('Utils', () => {
         });
     });
 
-    describe('withoutNil', () => {
-        let input: any;
-        let output: any;
-
-        beforeEach(() => {
-            input = {
-                a: 1,
-                b: null,
-                c: 0,
-                d: undefined,
-                e: {
-                    example: true,
-                    other: null,
-                }
-            };
-
-            output = withoutNil(input);
-        });
-
-        it('should copy the top level object', () => {
-            expect(output).not.toBe(input);
-        });
-
-        it('should not copy a nested object', () => {
-            expect(output.e).toBe(input.e);
-        });
-
-        it('should remove the nil values from the object', () => {
-            expect(output).toHaveProperty('a', 1);
-            expect(output).not.toHaveProperty('b');
-            expect(output).toHaveProperty('c', 0);
-            expect(output).not.toHaveProperty('d');
-            expect(output).toHaveProperty('e', {
-                example: true,
-                other: null,
-            });
-        });
-    });
-
     describe('parseList', () => {
         test.each([
             ['a', ['a']],
@@ -222,22 +181,6 @@ describe('Utils', () => {
             [[1], 1],
         ])('should convert %j to be %j', (input, expected) => {
             expect(toNumber(input)).toEqual(expected);
-        });
-    });
-
-    describe('toSafeString', () => {
-        test.each([
-            ['hello-there', 'hello-there'],
-            ['++_--hello', 'hello'],
-            ['Hello_Hi THERE', 'hello_hi_there'],
-            ['Howdy-Hi?+ you', 'howdy-hi-you'],
-            ['Hi there#*', 'hi-there'],
-            ['<HI> 3', 'hi-3'],
-            ['123', '123'],
-            [123, '123'],
-            [null, ''],
-        ])('should convert %s to be %s', (input, expected) => {
-            expect(toSafeString(input)).toEqual(expected);
         });
     });
 });
